Item Form and Controller Fixed
The Item form was not sending the categories and the status params
inside of the item hash so they were not being picked up by
strong params. Fixing this issue revealed some more problems within
our item creation and update methods that had to be dealt with as well.
I'm not sure why these did not come to light before but they are
working now.